change int to float on prices

+++ back/app/models/product_purchase.py
@@ -1,4 +1,4 @@
-from sqlalchemy import Column, Integer, ForeignKey
+from sqlalchemy import Column, Integer, ForeignKey, Float
 from sqlalchemy.orm import relationship
 from db.database import Base
 
@@ -10,7 +10,7 @@ class ProductPurchase(Base):
     purchase_id = Column(Integer, ForeignKey("purchases.id"))
     product_id = Column(Integer, ForeignKey("products.id"))
     quantity = Column(Integer)
-    unit_price = Column(Integer)
+    unit_price = Column(Float)
 
     purchase = relationship("Purchase", back_populates="product_purchases")
     product = relationship("Product", back_populates="product_purchases")
